requirements-pin.txt
matplotlib==3.8.0
numpy==1.25.2
pandas==1.5.3
scipy==1.11.3
mpi4py==3.1.5
scikit-learn==1.3.1
cupy-cuda11x==12.2.0

--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.1.0
--extra-index-url https://download.pytorch.org/whl/cu118
torchvision==0.16.0

--extra-index-url=https://pypi.nvidia.com
cuml-cu11==23.8.0
